added authorized user capabilities

+++ src/main/java/org/nastya/controller/AuthorController.java
@@ -3,6 +3,8 @@ package org.nastya.controller;
 import org.nastya.dto.AuthorFormDTO;
 import org.nastya.dto.AuthorListItemDTO;
 import org.nastya.service.AuthorService;
+import org.nastya.service.UserClient.UserClient;
+import org.nastya.service.UserClient.UserContext;
 import org.nastya.service.exception.AuthorNotFoundException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -21,9 +23,14 @@ import java.util.List;
 public class AuthorController {
     private static final Logger log = LoggerFactory.getLogger(AuthorController.class);
     private final AuthorService authorService;
+    private final UserClient userClient;
+    private final UserContext userContext;
 
-    public AuthorController(AuthorService authorService) {
+
+    public AuthorController(AuthorService authorService, UserClient userClient, UserContext userContext) {
         this.authorService = authorService;
+        this.userClient = userClient;
+        this.userContext = userContext;
     }
 
     @GetMapping
@@ -44,8 +51,11 @@ public class AuthorController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> deleteAuthor(@PathVariable int id, @RequestHeader("Session-ID") String sessionId) {
+    public ResponseEntity<String> deleteAuthor(@PathVariable int id, @RequestHeader("sessionId") String sessionId) {
         log.info("Deleting author by id '{}'", id);
+        if (!userClient.isUserAuthorized(userContext.getCurrentUserSession())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("User is not authorized to perform this action");
+        }
         try {
             authorService.deleteAuthor(id);
             log.info("Author '{}' deleted successfully", id);
@@ -57,8 +67,11 @@ public class AuthorController {
     }
 
     @PutMapping
-    public ResponseEntity<String> updateAuthor(@RequestBody AuthorFormDTO authorFormDTO, @RequestHeader("Session-ID") String sessionId) {
+    public ResponseEntity<String> updateAuthor(@RequestBody AuthorFormDTO authorFormDTO, @RequestHeader("sessionId") String sessionId) {
         log.info("Updating author with id '{}'", authorFormDTO.getId());
+        if (!userClient.isUserAuthorized(userContext.getCurrentUserSession())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("User is not authorized to perform this action");
+        }
         try {
             authorService.updateAuthor(authorFormDTO);
             log.info("Author '{}' updated successfully", authorFormDTO.getId());
@@ -70,8 +83,11 @@ public class AuthorController {
     }
 
     @PostMapping
-    public ResponseEntity<String> addAuthor(@RequestBody AuthorFormDTO authorFormDTO, @RequestHeader("Session-ID") String sessionId) {
+    public ResponseEntity<String> addAuthor(@RequestBody AuthorFormDTO authorFormDTO, @RequestHeader("sessionId") String sessionId) {
         log.info("Adding a new author");
+        if (!userClient.isUserAuthorized(userContext.getCurrentUserSession())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("User is not authorized to perform this action");
+        }
         try {
             authorService.addAuthor(authorFormDTO);
             log.info("Author '{}' added successfully", authorFormDTO.getId());
